requirements.txt
arabic-reshaper==2.0.14
matplotlib==2.0.0
nltk==3.4
numpy==1.14.3
opencv-python==3.3.1.11
pdf2image==1.4.1
pyinsane2==2.0.13
PySide2==5.11.2
tensorflow-gpu==1.13.1
python-docx==0.8.10
